Priority is now double, ShapeIntersection class

+++ Priority_Queue/PriorityQueueTree.h
@@ -2,6 +2,8 @@
 #include "PriorityQueueInterface.h"
 #include "AVLtree.h"
 
+using std::out_of_range;
+
 template <typename T>
 class PriorityQueueTree :
 	public PriorityQueueInterface <T>
@@ -12,18 +14,19 @@ public:
 	PriorityQueueTree()
 	{
 	}
-	PriorityQueueTree(T firstData, int priority)
+	PriorityQueueTree(T firstData, double priority)
 	{
 		PriorityTreeItem<T> item(firstData, priority);
 		tree.²nsert(item);
 	}
-	void Push(T data, int priority) override
+	void Push(T data, double priority) override
 	{
 		PriorityTreeItem<T> item(data, priority);
 		tree.²nsert(item);		
 	}
 	void Pop() override
 	{
+		if (Empty()) throw out_of_range("Atempt to pop from empty queue");
 		tree.DeleteKey(tree.MinNode()->data);
 	}
 	bool Empty() override
@@ -37,6 +40,7 @@ public:
 	}
 	T Top() override
 	{
+		if (Empty()) throw out_of_range("Queue is empty");
 		return tree.MinNode()->data.value;
 	}
 };
